Support hybrid dev+azure environments (#1892)
This change adds more support for 'hybrid' dev environments where
compute resources are deployed to the dev cluster, and non-compute
resources are deployed to azure.

The main contributions here are:

- Implementing a few more commonly-used ARM expression functions
- Tricky-tricks for resource ID handling.

The most significant thing is that we need to paper over the differences
in Radius running locally vs in Azure at the resource ID level. The IDs
have different meanings depending on whether the resource is a Radius
type or not. This can be simple again when we start using ARM
extensibility E2E for Radius.