Fix concurrent use of outputs
This change fixes some bad assumptions I had when writing the initial
version of this code, namely that concurrent writes to stdout are safe.
This isn't a guarantee Go provides.

The streams code is used when we're doing a build of multiple containers
in parallel. I noticed that `rad app deploy` could deadlock when one
container finishes significantly faster than the others which lead me to
finding this mistake.

The fix is to synchronize access to the output of the streams. I also
synchronized access to `NewStream` because it is also used concurrently.