Add optional azure credentials to Chart (#2084)
Part of: #2016

This change adds the optional ability to specify Azure credentials as
part of Helm Chart installation. These credentials are used by the RP to
authenticate with Azure.

A follow-up change will integrate these credentials with the Deployment
Engine and CLI for the provisioning flow.

Additionally, I added the ability to configure the kind of
authentication used as part of this change (optional). This means that
in the case of a misconfiguration, we'll get a clear error message.