Add AKS autoscaling for the node pool
Now that we're deploying the Radius control-plane on our kubernetes
cluster in Azure, we are seeing capacity problems causing test failures
due to pod limits.

This change adds autoscaling as part of the AKS configuration. We're
still creating AKS in a rather frugal configuration, but it will scale
up as needed while we run tests.