Improve error handling for invalid config
This change improves the error handling/reporting for cases where you
have an invalid CosmosDB configuration. Previously the error was a panic
with "Failed to create request headers: illegal base64 data at input
byte 3" as the only hint to the problem.

This change removes the panic (now it's just an error) as well as adding
more contextual information using wrapping along the failing code path.

Additionally I improved the logging behavior where panics get handled,
previously we were not logging here, and so there was no way to
determine the call stack the caused the failure.