Store Improvements (#2407)
* Use nested-tests to organize shared store tests

* Fix implementation of compareObjectLists

The test function for comparing objects was broken and wasn't actually
comparing anything.

* Implement "scope only" support for APIServer store

Fixes: #2371

This change implements the separation of 'resources' from 'scopes' in
the APIServer data store. This was previously implemented as part of the
ETCd store, but didn't have tests. This PR adds tests to the shared test
bed and implements the feature for APIServer.

* Implement support for query filters in store

Part of the work to unify store abstractions with the corerp.

This change implements client-side filtering for the ETCd and APIServer
stores based on the existing API.

* Update pkg/ucp/store/apiserverstore/apiserverclient.go