Don't specify plane as part of DE URL (#2697)
Found when investigating project-radius/deployment-engine#133, we need to remove /planes/radius/local from RADIUSBACKENDURL (which is the URL that the Deployment Engine will try to contact for the App Core RP) as that info will flow as part of scopes. This involves updating our helm chart to change the environment variable.