Redesign behavior of Scope queries (#2799)
Fixes: #2740

This change adjusts the behavior of Scope queries to be
more useful for our scenarios in UCP.

The scenarios that are impacted (where we had unuseful behaviors) are
related to queries of planes and resource groups for UCP. No behaviors
of the data store for our RPs are affected as they do not use Scope
queries.

The kinds of things done by our RPs are to "list <type> at <scope>" and
occaisionally to "list <type> at <scope> recursively". For example,
listing all `Applications.Core/containers` resources in a resource group
or subscription (recursive case). Translating the words ("list <type> at
<scope>") to our Query API has the desired effect for these cases.

The scope query API is unfortunately NOT symmetric, it does not
currently provide the same guarantees. Here's an example of an un-useful
behavior. Right now a query to "list all resources at
`/planes/radius/local`" would return `/planes/radius/local`

The current workaround of course is to "list all resources at <scope>
recursively". This would return the following:

```txt
/planes/radius/local
/planes/radius/local/resourceGroups/group1
/planes/radius/local/resourceGroups/group2
```

This is also unhelpful because because we don't expect
`/planes/radius/local` to be included.

The change here is to make the behavior symmetrical by storing data the
same way for scopes as we do for resources. We chop off the last segment
and treat that as the "effective routing scope" and store the previous
segments as the "effective root scope".