Slim down docker image for appcore rp (#2943)
Part of: #2582

This change switches us to distroless for the appcore rp docker image
from mariner. The overall impact is that the reduces the size from 245mb
to 53mb.

Additionally added ca-certificates to the UCP image. We will need these
to proxy to clouds like AWS and Azure, so adding them now.