Fix handling of different workspace cases
Fixes: #2953

Fixes: #2968

This change fixes the handling of different casing for workspace names
for #2953. Unfortunately viper lowercases all keys for storage, but it
**SILENTLY TOLERATES**  duplicates that differ by case. This means that
we need to be consistent in using lowercase everywhere we
programmatically interact with the workspace name. I did my best to add
verification for this, so we can avoid introducing more bugs in the
future.

For #2968 we had the wrong logic on a boolean check - the logic for
determining if a workspace was reusable was inverted. Oops.