Add setting to CLI for public endpoint override (#2991)
Part of: #2973

This change adds the `--public-endpoint-override` flag to CLI entry
points that install our helm chart. This way we can specify the publicly
accessible entry point of the cluster ingress for local testing, and
have it reflected in Radius commands.

Note that this change sets an existing setting that is supported by the
appcore-rp. We need other changes (URL on gateway) and support for `rad
app status` to test this end to end.