Fix Bicep Extensibility for Kubernetes
Fixes: #3077

This change fixes Bicep Extensibility for Kubernetes. The problem was a
simple mismatch of hostnames.

Most of the changes here were to enable an existing test in some kind of
reasonable form. The existing test for Kubernetes is blocked by
https://github.com/Azure/bicep/issues/7553 but with some minor
changes we can still verify the 'hello world' for Kubernetes, which we
need for one of our samples.

We'll also be fixing azure/bicep#7553 shortly.