Implement APIService contract (quietly!)
Fixes: #3091

This change implements two missing pieces of the Kubernetes APIService
contract that we previously left alone:

- Serving a discovery document
- Serving an OpenAPI v2 (swagger) document

Controller manager will ping you repeatedly for the discovery document
and api-server manager will ping you repeatedly for the openapi
document. This is the source of a TON of spam in our logs. Nothing was
really broken, but it was an annoying problem. For a user that closely
monitors metrics from a Kubernetes cluster this could show up in a way
that causes them to complain in the future. Prometheus once had a
similar issue.

Additionally, I implemented these handlers without logging anything
(deliberately). These are called pretty frequently by Kubernetes on the
happy-path, and for an 'idle' instance of UCP would generate a lot of
spam and push more interesting stuff below the fold.