Grant Bicep extensiblity permission on everything (#3165)
Fixes: #3160

This change grants Bicep extensibility maximal permissions on the
cluster. Right now we have a more limited set of permissions set which
lead to spotty limitations and failures.

This is ultimately something we'd like to review and come up with a more
nuanced policy. I'm adding this to our design queue to review.