Add support for scope collections to ID parser
This change adds support for parsing resource ids like:

```txt
/planes/radius/local/resourceGroups/my-group/resources
```

This is needed to support some of the additional functionality resource
groups provide, like listing the resources inside the group. We already
have this support for collections on resource types, this extends it to
cover scopes as well.

Unfortunately a few of our tests for 'invalid' IDs now fail because the
IDs they were using are now considered valid. I've updated those tests.