Consolidate rest.Result types
This change de-duplicates a lot of copy-pasted code between UCP and the
armrpc package. Standardizing on the armrpc versions of these types is
better because they are much more widely used right now and have tests.

There was one case where UCP was doing something specialized, so I added
it to armrpc and added a test.

The big win here will be the next change. Since we defined different
types for results we can't share controller code right now between
armrpc and ucp. That will be the next change.