De-duplicate controller interfaces
This change shares the controller interface between the ARMRPC package
and UCP. This should make it possible to use our generic controller
implementations inside UCP.

The main change here is that the UCP controllers require access to
http.ResponseWriter since they do proxying. I've added this to all of
the many many places where we have implemented controllers.

Next is to actually true using the shared controllers in UCP.