Simplify Kubernetes auth for server (#4233)
I happened upon this (old) code while working on something else, and
realized that it is still based on the CustomRP model. We have the
capability to connect to ARM and download credentials from AKS, which we
don't really need or want to do. The environment variables to configure
this are still present in the helm chart, but nothing sets them.

I've also learned a lot about how kubernetes configuration model works
since then, and this code can be simplified a lot. We don't really need
environment variables for configuration. If we're running in a cluster
then connecting to the cluster is the right thing (for now). So we can
remove the setting and simplify everyone's life.