Add location and tag support to Resource Groups
This change fixes bugs in the handling of location and tags for resource
groups. As-is these fields were not handled by the converter and so
would be omitted from the response even if they were specified on
PUT.

This change also fixes the output of the resource group list operation
to comply with the armrpc spec. Currently resource group list is
outputting a JSON array instead of an object. The swagger and generated
clients are correct, but appear to have never worked because the server
is wrong.

This change also specifies a location (my original goal) in all of the
places where create resource groups. Turns out tht we have a bug where
location is not validated #4429.

This change also adds these fields to the relevant tests and and does
some minor code cleanup. Many of these tests were using raw HTTP bodies
or alternate data-model types that allowed us to drift from what's
expected.