Add diagnostics for fallback, non-named workspaces
We're removing the dependency on workspaces that are stored in per-user
config in most commands to make the CLI more flexible. This is all
tracked as part of #4303.

This change adds errors and error messages for these cases as well as
the ability to detect them. My strategy is to start with
named-workspaces as a requirement for all of our commands and then
gradually roll out support for the more flexible cases. This way the PRs
stay incremental and reviewable.