Adds fallback workspace but block its use
This change adds the 'fallback workspace' concept to our configuration
system, but prevents its use in all of our commands. I'm taking this
strategy so that I can incrementally enable CLI functionality without
the need for a 'big bang' that touches all commands. I also realize that
this commit is a 'big bang' :) -- but I think that's justified because
each individual change is simple.

This change should have minimal impact on the UX for a Radius user, the
error message they will see if they have no workspace has changed
slightly, but no new functionality is enabled.

The commands that can be considered 'done' in this PR are the ones that
will **never** support fallback workspace for obvious reasons. For
instance `rad env switch` is a command that edits configuration, it will
never function without a configuration.