Add application scaffolding to rad init
Part of: #4303

This change adds 'application' scaffolding to `rad init`. When running
`rad init` the user has the option to treat the local directory as an
application. Doing so will output a sample `app.bicep` as well as a
configuration file that tells us the application name.

The intention is that this can be used in the root directory of a git
repo to configure the behavior of the CLI for the repo. This also will
support multiple applications per repo as well.

The next step here is that we'll integrate the CLI with the
configuration file so that our commands become aware of the application
name. We'll also pass the application id (when we know it) into Bicep as
a parameter so it can be used in Bicep files like the one we created.