Inject application parameter for deployment
This change adds the local directory config to all of our commands. This
means that the local directory config (`.rad/rad.yaml`) can now play the
roll of the `--application` parameter in most commands.

This also adds the application parameter to deployment. `rad deploy`
will inject the application resource id into Bicep when the parameter is
defined, and an application set as part of the command. This enables a
simpler authoring pattern where the application resource is not defined
in Bicep. We plan to show this to users first, and make this the default
for our samples and documentation.