Add `rad init --dev`
Part of: #4303

This change adds a `--dev` flag to `rad init` that is optimized for
development setup. The goal is for users to setup Radius for
non-production use a few scenarios:

- First time cluster setup
- Creating a new application for an existing cluster
- Accessing a cluster that's already setup by someone else

Passing `--dev` will short circuit **ALL** of the existing choices users
are asked today by `rad init` unless they absolutely need to be asked.
For example we'll choose the user's current KubeContext, default
namespace, default environment, no cloud providers, etc. We only ask
about these things if there's a conflict, for example if no environment
called 'default' exists.

We choosing an opt-in mechanism for this because we want to lean
*heavily* into optimizing for developer scenarios. It feels better to
provide a flexible option as the default and then have users opt-in to
non-production use in an obvious way. For an example, we'll include the
'dev' recipe pack by default without the user have to know what a
recipe pack is.