Add `rad init --dev` (#4510)
Part of: #4303

This change adds a `--dev` flag to `rad init` that is optimized for
development setup. The goal is make it trivial for users to setup Radius
for non-production use across a few scenarios:

- First time cluster setup
- Creating a new application for an existing cluster
- Accessing a cluster that's already setup by someone else

Passing `--dev` will short circuit **ALL** of the existing choices users
are asked today by `rad init` unless they absolutely need to be asked.
For example we'll choose the user's current KubeContext, default
namespace, default environment, no cloud providers, etc. We only ask
about these things if there's a conflict, for example if multiple
environments exist but none of them are called 'default'.

We choosing an opt-in mechanism for this because we want to lean
*heavily* into optimizing for developer scenarios. It feels better to
provide a flexible option as the default and then have users opt-in to
non-production use in an obvious way. For an example, we'll include the
'dev' recipe pack by default without the user have to know what a recipe
pack is. To make this absolutely clear - it's not just that we're
choosing defaults for some of our choices, it's that we're going to make
those defaults **heavily** biased towards non-production use. If we were
just choosing more defaults, we'd simplify further by baking it into
`rad init`.