Update deploy command to new framework
This change updates the deploy command to use the new CLI framework and
adds tests. I'm planning to make some changes to the deploy command, and
to re-use some of this infrastructure in another context as well. This
seemed like a good opportunity to add tests based on the current
behavior.