Remove more unused code
I noticed some more obsolete code while cleaning up some of our CLI
commands. All of this functionality are things we don't need at all, or
will bring back in a different form if we ever do.

The `testenv` command is a utility we used for setting up test
environments, or more recently running the DE locally as a process. We
don't have a scenario where the team uses it anymore, and it's not
documented anywhere so I think it's safe to remove.

The schemas folder contained the old swagger schemas for the CustomRP so
they are totally obsolete.