Add api version parameter to UCP credential resource (#4645)
# Description

UCP has now added API versioning support. This parameter is missing in
the swagger for the UCP credentials resource. Adding it.

#4609 

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: #4609 

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it