Add port-forwarding to `rad run`
Part of: #4303

This change adds support for port-forwarding to `rad run`. The previous
PR in this series added the port-forward infra and this change wires it
up to our command.

Since port-forwarding requires networking, I've added a new functional
test to cover the new functionality.

There's one cosmetic change here to the display based on feedback I got
from others. We want to show the pod name both for consistency with
stern and also so it can be copy-pasted if needed into kubectl commands.