Inject application parameter for deployment (#4638)
This change adds the local directory config to all of our commands. This
means that the local directory config (`.rad/rad.yaml`) can now play the
role of the `--application` parameter in most commands.

This also adds the application parameter to deployment. `rad deploy`
will inject the application resource id into Bicep when the parameter is
defined, and an application set as part of the command. This enables a
simpler authoring pattern where the application resource is not defined
in Bicep. We plan to show this to users first, and make this the default
for our samples and documentation.

The previous PR in this series enabled scaffolding of an `app.bicep`
that used the `application` parameter and avoided declaring the
application resource in Bicep. With this change, that scaffolded
`app.bicep` can be deployed without any further ceremony.