Support fallback workspace in environment commands (#4676)
This change adds support for the fallback workspace in the remaining
environment commands. This means that it's now possible to use commands
like `rad env show` without a workspace.

In these cases it will be required to specify the resource group. This
change tries to preserve most of the existing functionality and
semantics of the env commands. We may do more work in the future to make
them less reliant on resource groups.