Auto-create app as part of deployment
Part of: #4303

This change automatically creates the application resource during
deployment when an application is specified by the deployment command.

The app resource is only created if it does not already exists. This
means we'll never overwrite your app resource if you set it up in a
non-default way. This also means if you create the app resource in
Bicep, the default we set up will be immediately overwritten by your
version.

As I've been playing with `rad deploy -a` and `rad run` this feels like
a really necessary missing piece of the workflow. This way you can move
across different environments without any bootstrapping.