Add `rad run` command (#4702)
Part of: #4303

This change adds a `rad run` command. This is similar to `rad deploy`
except it continues running after deployment and displays streaming logs
for all of the application's pods.

`rad run` requires an application name to be passed in at the command
line, or specified via `rad.yaml`. We need to know the name of the
application in order to stream logs for it.

The implementation embeds and reuses `rad deploy`, and adds the unique
`rad run` functionality after running `rad deploy`'s logic. This allows
us to keep the two commands in sync.

Here's what it looks like:


![image](https://user-images.githubusercontent.com/1430011/203189929-7e668a77-c717-4f01-ba15-404533a9a24a.png)