Add application-scoped port-forward infra (#4769)
This change adds infrastructure for port-forwarding whole applications
at a time. This means we'll look for all of the pods and labels that
match our deployments, and port-forward ALL of them. The next step will
integrate this with `rad run`.

We have existing functionality for `rad resource expose` so the actual
heavy lifting with Kubernetes is based on existing code. I'll also add
e2e tests as part of the next step so that's covered. What's new here is
the plumbing to watch a set of deployments instead of just a single pod.

Some features of this that are relevant:

- We ignore 'stale' replicasets so that we always pick up the new pods
after doing a deployment (this is a common annoyance with k8s tools).
- We watch one pod-per-deployment max. We always choose the newest.
- We behave very gracefully during upgrades. When the pod we're
connected to **starts** its shutdown we will connect to a new one.
- We can include resources created by recipes as long as they are
labeled to match our expectations.
- We prefer to use the same local port as the container uses if
possible. If not possible, we will dynamically assign the port and print
it as part of the logs.
- There are no retries or reconnections right now. If there's demand for
this we will add it.

# Description

_Please explain the changes you've made._

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: #issue_number

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it