Add port-forwarding to `rad run` (#4800)
Part of: #4303

This change adds support for port-forwarding to `rad run`. The previous
PR in this series added the port-forward infra and this change wires it
up to our command.

Since port-forwarding requires networking, I've added a new functional
test to cover the new functionality.

There's one cosmetic change here to the display based on feedback I got
from others. We want to show the pod name both for consistency with
stern and also so it can be copy-pasted if needed into kubectl commands.

# Description

_Please explain the changes you've made._

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: #issue_number

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it