Fix datarace in MaxConcurrency Test  (#4855)
# Description

I am unable to reproduce this datarace problem from my test. Based on
the error messages in #4781 , datarace happened when
`updateResourceState` update map data which is returned from mock. In
real world, this mocked map data should be the different instance, but
in the test, mocked Get function returned the same instance. So there
are the possibility to update map data concurrently.

The fix will return new object instance to avoid datarace in test when
mocked function is called.

## Issue reference

<!--
We strive to have all PR being opened based on an issue, where the
problem or feature have been discussed prior to implementation.
-->

Fixes: #4781 

## Checklist

Please make sure you've completed the relevant tasks for this PR, out of
the following list:

* [ ] Code compiles correctly
* [ ] Adds necessary unit tests for change
* [ ] Adds necessary E2E tests for change
* [ ] Unit tests passing
* [ ] Extended the documentation / Created issue for it