Disable K8s Service Links (#4870)
Fixes: #3002

This change disables Kubernetes service links, a feature that injects
environment variables for each service into pods. We're disabling this
feature because it has plenty of issues and it's platform-specific
(kubernetes). See the issues mentioned in comments for context.

Radius provides its own service-discovery mechanism without the same
problems (connections).