Use generic GET/LIST controllers for Redis cache
This is a pilot for adopting the generic controllers for the link RP.
The first PR just includes Redis and serves as an example of the changes
we'll need to make to make link RP compatible with the generic
controllers.

Once this is approved, I'll switch the rest of the link RP to use the
generic GET/LIST controllers.